:recycle: refactor: call portfolios-tab-content and accounts-tab-content directly, instead use another file to wrap them
